move LocalStoreInvoice to production
