+++ test/invoice/local_store_invoice_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hive_flutter/adapters.dart';
 import 'package:thepos/features/invoice/models/invoice.dart';
+import 'package:thepos/features/invoice/repositories/local_invoice/local_store_invoice.dart';
 
 import 'helpers/shared_test_helper.dart';
 
@@ -48,26 +49,4 @@ void main() {
     final List<Invoice> result = localStoreInvoice.retrieve();
     expect(result.isEmpty, true);
   });
-}
-
-class LocalStoreInvoice {
-  LocalStoreInvoice({required this.hiveBox});
-
-  final Box<Map<String, dynamic>> hiveBox;
-
-  Future<Invoice> store(Invoice invoice) async {
-    await hiveBox.put(invoice.clientId, invoice.toJson());
-    return invoice;
-  }
-
-  List<Invoice> retrieve() {
-    return hiveBox.values
-        .map((Map<String, dynamic> json) => Invoice.fromJson(json))
-        .toList();
-  }
-
-  Future<Invoice> delete(Invoice invoice) async {
-    await hiveBox.delete(invoice.clientId);
-    return invoice;
-  }
 }
